requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.0
certifi==2025.4.26
charset-normalizer==3.4.0
contourpy==1.3.2
cycler==0.12.1
decorator==5.1.1
deepspeed==0.15.2
defusedxml==0.7.1
diffuser==0.0.1
diffusers==0.31.0
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2024.10.0
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.67.0
h5py==3.12.1
hf-xet==1.1.2
hjson==3.1.0
huggingface-hub==0.32.2
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
libclang==18.1.1
lightly==1.5.20
lightly-utils==0.0.2
lightning-utilities==0.14.3
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.4.2
ninja==1.11.1.1
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.13.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.13.2
pickleshare==0.7.5
pillow==11.0.0
pip==25.1.1
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.3.1
protobuf==3.19.6
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.14.1
segmentation_models_pytorch==0.5.0
setuptools==80.8.0
simsimd==6.2.1
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
stringzilla==3.12.5
sympy==1.14.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.11
tinycss2==1.3.0
tokenizers==0.21.1
torch==2.7.0
torchao==0.6.1
torchmetrics==1.7.2
torchvision==0.22.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.52.4
triton==3.3.0
typing_extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
wheel==0.45.1
wrapt==1.16.0
yarg==0.1.9
yarl==1.20.0
zipp==3.20.2
